+++ saleor/graphql/product/filters.py
@@ -36,7 +36,7 @@ from ...product.models import (
 from ...product.search import search_products
 from ...warehouse.models import Allocation, Reservation, Stock, Warehouse
 from ..channel.filters import get_channel_slug_from_filter_data
-from ..core.descriptions import ADDED_IN_38, ADDED_IN_316
+from ..core.descriptions import ADDED_IN_38, ADDED_IN_317
 from ..core.doc_category import DOC_CATEGORY_PRODUCTS
 from ..core.filters import (
     BooleanWhereFilter,
@@ -1261,7 +1261,7 @@ class CategoryFilter(MetadataFilterBase):
     updated_at = ObjectTypeFilter(
         input_class=DateTimeRangeInput,
         method=filter_updated_at_range,
-        help_text=f"Filter by when was the most recent update. {ADDED_IN_316}",
+        help_text=f"Filter by when was the most recent update. {ADDED_IN_317}",
     )
 
     class Meta: